pyproject.toml
[project]
name = "sdd-fastapi-reference"
version = "0.1.0"
description = "User Authentication API with FastAPI, SQLModel, and JWT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.16",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "coverage>=7.4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
